UPSTREAM: sepolicy: Add CAP_WAKE_ALARM to system_server.te
With v4.8+ kernels, CAP_WAKE_ALARM is needed to set
alarmtimers via timerfd (this change is likely to be
backported to stable as well).

However, with selinux enabled, we also need to allow
the capability on the system_server so this enables it.

Change-Id: I7cd64d587906f3fbc8a129d48a4db07373c74c7e
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from commit 19b6485f5ed005f1ae9b5ba06b3baca4639a4c83)